Move reporter benchmarking code to bench/reporter
